requirements.txt
Django
djangorestframework
django-cors-headers
pyjwt==2.1.0